feat: impl verify
