+++ crates/sdk/src/utils.rs
@@ -8,12 +8,6 @@ use std::{
     time::Duration,
 };
 
-use std::{
-    sync::Once,
-    thread::{sleep, spawn},
-    time::Duration,
-};
-
 use sp1_core_machine::io::SP1Stdin;
 pub use sp1_core_machine::utils::setup_logger;
 use sysinfo::{MemoryRefreshKind, RefreshKind, System};
@@ -85,63 +79,4 @@ pub(crate) fn block_on<T>(fut: impl std::future::Future<Output = T>) -> T {
 pub(crate) fn check_release_build() {
     #[cfg(debug_assertions)]
     panic!("sp1-sdk must be built in release mode, please compile with the --release flag.");
-}
-
-/// The cycle limit and gas limit are determined according to the following priority:
-///
-/// 1. If either of the limits are explicitly set by the requester, use the specified value.
-/// 2. If simulation is enabled, calculate the limits by simulating the execution of the program.
-///    This is the default behavior.
-/// 3. Otherwise, use the default limits ([`DEFAULT_CYCLE_LIMIT`] and [`DEFAULT_GAS_LIMIT`]).
-#[cfg(feature = "network")]
-pub(crate) fn get_execution_limits<C: SP1ProverComponents>(
-    prover: &SP1Prover<C>,
-    cycle_limit: Option<u64>,
-    gas_limit: Option<u64>,
-    elf: &[u8],
-    stdin: &SP1Stdin,
-    skip_simulation: bool,
-) -> Result<(u64, u64, Option<Vec<u8>>)> {
-    use crate::network::{Error, DEFAULT_CYCLE_LIMIT, DEFAULT_GAS_LIMIT};
-
-    let cycle_limit_value = if let Some(cycles) = cycle_limit {
-        cycles
-    } else if skip_simulation {
-        DEFAULT_CYCLE_LIMIT
-    } else {
-        // Will be calculated through simulation.
-        0
-    };
-
-    let gas_limit_value = if let Some(gas) = gas_limit {
-        gas
-    } else if skip_simulation {
-        DEFAULT_GAS_LIMIT
-    } else {
-        // Will be calculated through simulation.
-        0
-    };
-
-    // If both limits were explicitly provided or skip_simulation is true, return immediately.
-    if (cycle_limit.is_some() && gas_limit.is_some()) || skip_simulation {
-        return Ok((cycle_limit_value, gas_limit_value, None));
-    }
-
-    // One of the limits were not provided and simulation is not skipped, so simulate to get one
-    // or both limits
-    let execute_result = prover
-        .execute(elf, stdin, sp1_core_executor::SP1Context::builder().calculate_gas(true).build())
-        .map_err(|_| Error::SimulationFailed)?;
-
-    let (_, committed_value_digest, report) = execute_result;
-
-    // Use simulated values for the ones that are not explicitly provided.
-    let final_cycle_limit =
-        if cycle_limit.is_none() { report.total_instruction_count() } else { cycle_limit_value };
-    let final_gas_limit =
-        if gas_limit.is_none() { report.gas.unwrap_or(DEFAULT_GAS_LIMIT) } else { gas_limit_value };
-
-    let public_values_hash = Some(committed_value_digest.to_vec());
-
-    Ok((final_cycle_limit, final_gas_limit, public_values_hash))
 }
